requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
authlib
itsdangerous
httpx
python-dotenv
mysql-connector-python
pymysql
sqlalchemy
alembic
jinja2
pydantic[email]
requests